refactoring: fix socket io client wrapper, apply config
